requirements.txt
PyQt6
PyPDF2
Pillow
pytesseract